real_estate_backend/requirements.txt
django==3.2
django-appconf==1.0.5
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.2.1
django-filter==23.2
django-kn-defaults==0.3.1
django-oauth2-authcodeflow==0.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
pytest-cov==3.0.0
flake8==4.0.1
flake8-isort==4.1.1
black==21.12b0
pylint-django==2.5.0
pre-commit==2.16.0
factory-boy==3.2.1
djvue==0.2.1
pillow==8.3.1
django-cors
python-dotenv==0.20.0
django-cors-headers==3.7.0
psycopg2-binary
bcrypt==4.0.1
drf-spectacular==0.26.2
boto3
google==3.0.0
google-auth==2.19.1
google-cloud==0.34.0